requirements.txt
flask
pandas
numpy
scikit-learn
google-generativeai
python-dotenv
requests
